Update 1WhyAcceptDoge
The goal is to create an "At a first glance" flyer that briefly illustrates to small/local businesses the the benefits for them to accepting Dogecoin in a easy to understand and eye catching format. We want to keep this VERY simple, yet sufficiently informative. Instead of simply asking "Can I pay with Dogecoin?" and leaving the employee wondering what to answer, this pamphlet's goal is to create an opportunity for the Business Owners to be able to learn more about how they can benefit from using Dogecoin at their own time of convenience and be able to find out more or reach back for support in case they are interested.