Update How2AcceptDoge
Create a more detailed pamphlet with more detailed information, but still something that can be printer and passed around. It should include resources and instructions on how to set up a wallet and start being able to send/receive Dogecoin. Definitely a QR code linking to the website.
It would be great to pair this pamphlet with a window sticker that says something like "Dogecoin Accpeted Here", and definitely with a Much Wow Shiba Inu Dog on it. :-)