Update 3KnowTheDogeWebsite
Creating a website with information about different aspects of dogecoin, various wallet options, security tips, and (possibly most importantly) a step by step guide to getting started with a dogecoin wallet or a dogecoin payment system (I'd love to be able to set online businesses with a version of Tesla's merch store Doge Payment Engine). Priority on wallets with the lowest transaction fee, fastest processing times (showing the txn with pending status would be sufficent I believe for a transaction to be considered complete) and best security possible.